pyproject.toml
[project]
name = "library"
version = "1.0.0"
authors = [
    {name = "Anmol Sharma", email = "anmolsharma009a@gmail.com"}
]
description = "A Library Management System API built with Flask having tokem based authentication."
dependencies = [
    "flask",
    "pyjwt",
]
readme = "README.md"
requires-python = ">=3.10"

[build-system]
requires = ["flit_core>=3.4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["library"]